Bug where song restarts when paused fixed

+++ app/src/main/java/com/ece493/group5/adjustableaudio/ui/media_player/MediaPlayerFragment.java
@@ -379,7 +379,7 @@ public class MediaPlayerFragment extends Fragment
             public void onProgressChanged(SeekBar seekBar, int i, boolean b)
             {
                 //TODO: update UI time display
-                songSeekBarPosition = (int) i;
+                songSeekBarPosition = i;
             }
 
             @Override
